added README and cleanup
