+++ scripts/gmshlayerbuilder/__main__.py
@@ -1,13 +1,13 @@
 from argparse import ArgumentParser
 
 try:
-    import _gmshlayerbuilder
+    import _gmshlayerbuilder  # noqa: F401
 except ImportError:
     import sys
     import os
 
     sys.path.append(os.path.dirname(__file__))
-    import _gmshlayerbuilder
+    import _gmshlayerbuilder  # noqa: F401
 
 
 def get_parser():
@@ -37,18 +37,22 @@ def get_parser():
     return parser
 
 
-def run():
+def run2D():
+    import _gmshlayerbuilder.dim2
+
     args = get_parser().parse_args()
-    builder = _gmshlayerbuilder.dim2.topo_reader.builder_from_topo_file(args.topo_file)
+    builder = _gmshlayerbuilder.dim2.layer_builder.topo_reader.builder_from_topo_file(
+        args.topo_file
+    )
 
     model = builder.create_model()
     if args.should_plot:
-        _gmshlayerbuilder.dim2.plotter.plot_model(model)
+        _gmshlayerbuilder.dim2.layer_builder.plotter.plot_model(model)
 
-    _gmshlayerbuilder.dim2.exporter.Exporter2D(
+    _gmshlayerbuilder.dim2.layer_builder.exporter.Exporter2D(
         model, args.output_folder, nonconforming_adjacencies_file="nc_adjacencies"
     ).export_mesh()
 
 
 if __name__ == "__main__":
-    run()
+    run2D()
