--- scripts/gmshlayerbuilder/_gmshlayerbuilder/__init__.py
@@ -1,3 +1,0 @@
-from . import dim2
-
-__all__ = ["dim2"]
