+++ scripts/gmshlayerbuilder/_gmshlayerbuilder/dim2/__init__.py
@@ -1,6 +1,6 @@
-from .layer2d import LayeredBuilder
+from . import layer_builder
+from .exporter import Exporter2D
 from .topo_reader import builder_from_topo_file
 from .plotter import plot_model
-from .exporter import Exporter2D
 
-__all__ = ["LayeredBuilder", "builder_from_topo_file", "plot_model", "Exporter2D"]
+__all__ = ["layer_builder", "Exporter2D", "builder_from_topo_file", "plot_model"]
